% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{benchmark}
\alias{benchmark}
\title{microbenchmark selected}
\usage{
benchmark(runs = 10)
}
\description{
microbenchmark selected code for 10 runs in console without changing the
source code.
}
\details{
\code{microbenchmark()} parameters can be changed by recalling history in
console then editing the last line.
}
