% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clip_2_df}
\alias{clip_2_df}
\title{read clipboard into data frame}
\usage{
clip_2_df()
}
\value{
the data frame
}
\description{
Read windows/mac/linux clipboard, convert csv into data frame, open data
viewer, return the data frame
}
