% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clip_read_lines}
\alias{clip_read_lines}
\title{Read windows/mac clipboard into lines}
\usage{
clip_read_lines()
}
\value{
character vector
}
\description{
If windows, call \code{utils::readClipboard()}. If mac os, use
\code{pipe("pbpaste")}.
}
\examples{
clip_read_lines()
}
