% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clip_write_lines}
\alias{clip_write_lines}
\title{Write lines into windows/mac clipboard}
\usage{
clip_write_lines(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
If windows, call \code{utils::writeClipboard()}. If mac os, use
\code{pipe("pbcopy", "w")}.
}

