% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{flip_windows_path}
\alias{flip_windows_path}
\title{Flip windows path}
\usage{
flip_windows_path()
}
\description{
\code{flip_windows_path} convert "\" in clipboard to "/", then paste into
current cursor position
}

