% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_console}
\alias{format_console}
\title{Convert Console Print Out to Script}
\usage{
format_console()
}
\description{
Read console input and output from clipboard, format as script(remove the >
prompt, convert output as comments).
}
\details{
Formated script is written back to clipboard, and inserted to current cursor
location
}
