% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{printc}
\alias{printc}
\title{Generate the literal c() of a character vector}
\usage{
printc(x)
}
\arguments{
\item{x}{a vector holding items x1, x2, ...}
}
\value{
string of "c("x1", "x2")"
}
\description{
It's often needed to create a vector literally even the vector can be
obtained in code. For example a sbuset of column names can be get with number
index, but it's not safe to use number index in code. Use this function to
turn the output of number index into literal format, then you can copy the
output to code.
}
\examples{
printc(names(mtcars)[1:3])

}
