% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{profv}
\alias{profv}
\title{profvis selected}
\usage{
profv()
}
\description{
profvis selected code in console without changing source code. RStudio have a
similar builtin menu in editor toolbar, but that only works with .R script,
not working in .Rmd or unsaved file.
}
