% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_render.R
\name{render_addin}
\alias{render_addin}
\title{View obj/variable \code{str()} output in html view by addin}
\usage{
render_addin()
}
\description{
View obj/variable \code{str()} output in html view by addin
}
