% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{render_rmd}
\alias{render_rmd}
\title{Render RMarkdown in global environment}
\usage{
render_rmd()
}
\description{
Knit document will start from scratch, this will use global environment
instead. So you don't have to run some expensive operations like read huge
file from disk every time in rendering.
}
