% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_render.R
\name{render_str}
\alias{render_str}
\title{Render str() output in html with style}
\usage{
render_str(obj)
}
\arguments{
\item{obj}{the object/variable to be inspected}
}
\description{
If the output of \code{str()} is long, user can easily get lost in scrolling.
This function convert output into html with tag classes and render it with
style. You can customize the style by experimenting with chrome/firefox
developer tools first, then save to css. The css is located in
\code{system.file("css", "str_render.css", package = "mischelper")}
}
\details{
By default a browser window will be opened because it is intended for long
output.
}

