% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scan_fun}
\alias{scan_fun}
\title{Scan Function Object For External Functions}
\usage{
scan_fun(fun_obj)
}
\arguments{
\item{fun_obj}{}
}
\value{
result table
}
\description{
Scan Function Object For External Functions
}
\examples{
mischelper::scan_fun(sort)
}
