% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{time_it}
\alias{time_it}
\title{time selected code}
\usage{
time_it()
}
\description{
Easier way to measure time cost of expression or selected code. Underthehood
it's just microbenchmark run once instead of 10.
}
