% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap text}
\usage{
unwrap(extra_blank_line = FALSE)
}
\arguments{
\item{extra_blank_line}{Whether to insert extra blank line between paragraphs.}
}
\description{
\code{unwrap} Remove unneeded hard line breaks of text in clipboard, then
paste into current cursor position.
}
