% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formater.R
\name{unwrap_extra_blank}
\alias{unwrap_extra_blank}
\title{Unwrap with blank line}
\usage{
unwrap_extra_blank()
}
\description{
Remove unneeded hard line breaks of text in clipboard, insert extra blank
line between paragraphs, then paste into current cursor position. Need this
helper because RStudio Addin doesn't support function with parameters.
}

