% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{benchmark}
\alias{benchmark}
\title{microbenchmark selected}
\usage{
benchmark()
}
\description{
microbenchmark selected code 10 runs in console without changing the source
code. microbenchmark parameters can be changed by recalling history in
console then edit the last line.
}

