% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_console}
\alias{format_console}
\title{Format console}
\usage{
format_console()
}
\description{
read console input and output from clipboard, format as script
}
\details{
Formated script is written back to clipboard, and inserted to current cursor
location
}
