% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMD_utils.R, R/misc.R
\name{printc}
\alias{printc}
\title{Print vector literal code for a string Vector}
\usage{
printc(x)

printc(x)
}
\arguments{
\item{x}{a vector holding items x1, x2, ...}
}
\value{
string of "c("x1", "x2")"
}
\description{
Sometimes we need to write a vector which values come from data, which means
to manually copy values and write into a vector definition. This function
will take a vector (often came from data, like all unique values of a
category column) and print a vector definition string, then it can be copied
to source code.

It's often needed to create a vector literally even the vector can be
obtained in code. For example a sbuset of column names can be get with number
index, but it's not safe to use number index in code. Use this function to
turn the output of number index into literal format, then you can copy the
output to code.
}
\examples{
printc(letters[1:3])
printc(names(mtcars)[1:3])

}
